Refactor solver to use iterative, explicit stack DFS
This had an unintended side-effect of SUPER EFFICIENTLY solving the maze because
it accidentally priortized DOWNWARD movement.

An artifact of the `Maze::get_adjacent_spaces` method is that it's always
returning `[left, right, above, below]` ordering. The recursive approach stepped
through directions from the beginning of the list, while the iterative approach
using a stack stepped through directions from the end.

The trick with these mazes is apparently to always move down when you can.