Replace floor division of rows/columns counting with rounded int div
Because of variability in the dimensions of images due to things like bad crops,
I previously had to undercount the size of the cells to get an approximation of
the max number of cells that could fit in the image; if 9.9 cells fit in a
dimension, the parse would thing it was only 9 cells instead of the more
reasonabl 10 cells where either end was slightly cut-off.

Now, it does rounded integer division to calculate approximately how many cells
could fit in each dimension.